fix(executor,planner): array_agg keeps NULLs; abs/mod/split_part PG-correct types + values
- array_agg now preserves NULL elements in both the hash and sort
  accumulation paths, matching PostgreSQL (string_agg still skips NULLs).
- abs() handles float4/float8/numeric/money and preserves the argument's
  own type (integer width included); the binder declares the matching
  type instead of always Int64.
- split_part() with an empty delimiter returns the whole string for
  field 1 and '' otherwise, matching PostgreSQL.
- mod() gains an integer fast path computing checked_rem on i64 (exact
  for values past f64's 53-bit mantissa) and returning the wider integer
  type; the binder declares integer for integer inputs, float8 otherwise.

+++ crates/ultrasql-executor/src/eval/tests/tests_column_literal.rs
@@ -178,7 +178,7 @@ fn cast_and_size_helpers_cover_oid_reg_and_text_surfaces() {
 
 #[test]
 fn text_math_regex_and_format_helpers_cover_common_scalar_paths() {
-    assert_eq!(eval_fn("abs", vec![Value::Int32(-7)]), Value::Int64(7));
+    assert_eq!(eval_fn("abs", vec![Value::Int32(-7)]), Value::Int32(7));
     assert_eq!(
         eval_fn("lower", vec![Value::Text("MiXeD".to_owned())]),
         Value::Text("mixed".to_owned())
